Refactor the server code into more files, and create a running loop on the server
