+++ Weave_Server/Weave_Server/include/Config.h
@@ -1,0 +1,12 @@
+#pragma once
+
+// Define that we are the server to help with dependency management
+#define WEAVE_SERVER
+
+#define DEF_LISTEN_PORT     50000
+#define DEF_RESPONSE_PORT   50001
+#define DEF_BUF_SIZE        512
+
+// Command line args
+#define HELP_FLAG              "-h"
+#define CONFIG_FILE_FLAG       "-c"
